requirements.txt
# Web Framework
fastapi==0.134.0
uvicorn[standard]==0.41.0
python-multipart==0.0.22

# LangChain & LangGraph (1.x)
langgraph>=1.0.0
langgraph-checkpoint>=2.0.0
langgraph-checkpoint-postgres>=2.0.0
langchain>=1.0.0
langchain-openai>=1.0.0
langchain-anthropic>=1.0.0
langchain-core>=1.0.0
langchain-community>=0.4.0
langsmith>=0.4.0
mem0ai>=0.0.7

# Database & Persistence
psycopg[binary]==3.3.3
pgvector==0.4.2
sqlalchemy>=2.0.25

# Tools
tavily-python==0.7.22
e2b-code-interpreter==2.4.1
playwright>=1.47.0
langchain-mcp-adapters==0.2.1
mcp==1.26.0
dashscope>=1.24.6  # DashScope SDK (ASR/TTS)

# Utilities
python-dotenv==1.2.1
pydantic==2.12.5
pydantic-settings==2.13.1
httpx==0.28.1

# Optional heavy tool deps moved to requirements-optional.txt
# Async support
aiofiles>=24.1.0
prometheus-client==0.24.1
tiktoken>=0.6.0
pillow>=10.0.0
requests>=2.31.0
markdownify>=0.12.0
matplotlib>=3.9.0
greenlet>=3.1.1

# RAG & Export (optional but commonly used)
# These have try/except guards - install for full functionality
pymupdf>=1.23.0           # PDF parsing
python-docx>=0.8.11       # DOCX parsing/export
chromadb>=0.4.0           # Vector store for RAG
weasyprint>=60.0          # PDF export
jinja2>=3.1.0             # HTML template rendering
markdown>=3.5.0           # Markdown to HTML conversion
openai>=1.0.0             # OpenAI embeddings API
minio>=7.2.0             # MinIO object storage
pymilvus>=2.4.0          # Milvus vector database

# Real-time Feed Providers (optional)
# Install for Twitter/Reddit/HackerNews search
tweepy>=4.14.0            # Twitter/X API v2
praw>=7.7.0               # Reddit API
ddgs>=9.0.0               # DuckDuckGo (no API key)

# Academic Search Providers (optional)
# Install for arXiv/Semantic Scholar/PubMed search
arxiv>=2.1.0              # arXiv API
biopython>=1.83           # PubMed/NCBI Entrez
exa-py>=1.0.0             # Exa.ai neural search
